+++ hutool-core/src/main/java/cn/hutool/core/net/DefaultTrustManager.java
@@ -1,12 +1,13 @@
-package cn.hutool.http.ssl;
+package cn.hutool.core.net;
 
 import javax.net.ssl.X509TrustManager;
 import java.security.cert.X509Certificate;
 
 /**
- * 证书管理
+ * 默认信任管理器，默认信任所有客户端和服务端证书
  *
  * @author Looly
+ * @since 5.5.7
  */
 public class DefaultTrustManager implements X509TrustManager {