+++ hutool-extra/src/main/java/cn/hutool/extra/mail/MailUtil.java
@@ -5,6 +5,8 @@ import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.map.MapUtil;
 import cn.hutool.core.util.StrUtil;
 
+import javax.mail.Authenticator;
+import javax.mail.Session;
 import java.io.File;
 import java.io.InputStream;
 import java.util.Collection;
@@ -344,6 +346,24 @@ public class MailUtil {
 	public static String send(MailAccount mailAccount, Collection<String> tos, Collection<String> ccs, Collection<String> bccs, String subject, String content, Map<String, InputStream> imageMap,
 	                          boolean isHtml, File... files) {
 		return send(mailAccount, false, tos, ccs, bccs, subject, content, imageMap, isHtml, files);
+	}
+
+	/**
+	 * 根据配置文件，获取邮件客户端会话
+	 *
+	 * @param mailAccount 邮件账户配置
+	 * @param isSingleton 是否单例（全局共享会话）
+	 * @return {@link Session}
+	 * @since 5.5.7
+	 */
+	public static Session getSession(MailAccount mailAccount, boolean isSingleton){
+		Authenticator authenticator = null;
+		if (mailAccount.isAuth()) {
+			authenticator = new UserPassAuthenticator(mailAccount.getUser(), mailAccount.getPass());
+		}
+
+		return isSingleton ? Session.getDefaultInstance(mailAccount.getSmtpProps(), authenticator) //
+				: Session.getInstance(mailAccount.getSmtpProps(), authenticator);
 	}
 
 	// ------------------------------------------------------------------------------------------------------------------------ Private method start