+++ hutool-http/src/main/java/cn/hutool/http/server/action/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * {@link com.sun.net.httpserver.HttpServer} 封装
+ *
+ * @author looly
+ */
+package cn.hutool.http.server.action;