+++ hutool-json/src/main/java/cn/hutool/json/JSONString.java
@@ -1,8 +1,8 @@
 package cn.hutool.json;
 
 /**
- * <code>JSONString</code>接口定义了一个<code>toJSONString()</code><br>
- * 实现此接口的类可以通过实现<code>toJSONString()</code>方法来改变转JSON字符串的方式。
+ * {@code JSONString}接口定义了一个{@code toJSONString()}<br>
+ * 实现此接口的类可以通过实现{@code toJSONString()}方法来改变转JSON字符串的方式。
  * 
  * @author Looly
  *