+++ CHANGELOG.md
@@ -3,7 +3,52 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.5.5 (2020-12-24)
+# 5.5.8 (2021-01-09)
+
+### 新特性
+### Bug修复
+* 【core   】     修复FileUtil.move以及PathUtil.copy等无法自动创建父目录的问题（issue#I2CKTI@Gitee）
+
+-------------------------------------------------------------------------------------------------------------
+
+# 5.5.7 (2021-01-07)
+
+### 新特性
+* 【core   】     DynaBean.create增加重载方法（pr#245@Gitee）
+* 【core   】     IdcardUtil增加重载是否忽略大小写（issue#1348@Github）
+* 【poi    】     SheetRidReader增加getRidByIndex方法（issue#1342@Github）
+* 【extra  】     MailAccount增加sslProtocols配置项（issue#IZN95@Gitee）
+* 【extra  】     MailUtil增加getSession方法
+* 【setting】     新增setByGroup和putByGroup，set和put标记为过期（issue#I2C42H@Gitee）
+* 【crypto 】     修改SymmetricAlgorithm注释（issue#1360@Github）
+* 【all    】     pom中将META-INF/maven下全部exclude（pr#1355@Github）
+* 【http   】     SimpleServer中增加addFilter等方法，并使用全局线程池
+* 【core   】     CollUtil.forEach 增加null 判断（pr#250@Gitee）
+* 【extra  】     FtpConfig增加serverLanguageCode和systemKey配置,Ftp.download增加重载（pr#248@Gitee）
+
+### Bug修复
+* 【core   】     修复CsvReader读取双引号未转义问题（issue#I2BMP1@Gitee）
+* 【json   】     JSONUtil.parse修复config无效问题（issue#1363@Github）
+* 【http   】     修复SimpleServer返回响应内容Content-Length不正确的问题（issue#1358@Github）
+* 【http   】     修复Https请求部分环境下报证书验证异常问题（issue#I2C1BZ@Gitee）
+
+-------------------------------------------------------------------------------------------------------------
+
+# 5.5.6 (2020-12-29)
+
+### 新特性
+* 【core   】     手机号工具类 座机正则表达式统一管理（pr#243@Gitee）
+* 【extra  】     Mail增加setDebugOutput方法（issue#1335@Gitee）
+
+### Bug修复
+* 【core   】     修复ZipUtil.unzip从流解压关闭问题（issue#I2B0S1@Gitee）
+* 【poi    】     修复Excel07Writer写出表格错乱问题（issue#I2B57B@Gitee）
+* 【poi    】     修复SheetRidReader读取字段错误问题（issue#1342@Github）
+* 【core   】     修复FileUtil.getMimeType不支持css和js（issue#1341@Github）
+
+-------------------------------------------------------------------------------------------------------------
+
+# 5.5.5 (2020-12-27)
 
 ### 新特性
 * 【core   】     URLUtil.normalize新增重载（pr#233@Gitee）
@@ -13,14 +58,20 @@
 * 【poi    】     增加ExcelDateUtil更多日期格式支持（issue#1316@Github）
 * 【core   】     NumberUtil.toBigDecimal支持各类数字格式，如1,234.56等（issue#1334@Github）
 * 【core   】     NumberUtil增加parseXXX方法（issue#1334@Github）
+* 【poi    】     Excel07SaxReader支持通过sheetName读取（issue#I2AOSE@Gitee）
 
 ### Bug修复
 * 【core   】     FileUtil.isSub相对路径判断问题（pr#1315@Github）
 * 【core   】     TreeUtil增加空判定（issue#I2ACCW@Gitee）
 * 【db     】     解决Hive获取表名失败问题（issue#I2AGLU@Gitee）
 * 【core   】     修复DateUtil.parse未使用严格模式导致结果不正常的问题（issue#1332@Github）
+* 【core   】     修复RuntimeUtil.getUsableMemory非static问题（issue#I2AQ2M@Gitee）
+* 【core   】     修复ArrayUtil.equals方法严格判断问题（issue#I2AO8B@Gitee）
+* 【poi    】     修复SheetRidReader在获取rid时读取错误问题（issue#I2AOQW@Gitee）
+* 【core   】     修复强依赖了POI的问题（issue#1336@Github）
 
 -------------------------------------------------------------------------------------------------------------
+
 # 5.5.4 (2020-12-16)
 
 ### 新特性